+++ angular-workspace/projects/example-client-app/src/app/app.module.ts
@@ -7,7 +7,8 @@ import { NimbleTextAreaModule, NimbleTextFieldModule, NimbleNumberFieldModule, N
     NimbleTabModule, NimbleTabPanelModule, NimbleTabsModule, NimbleTabsToolbarModule, NimbleMenuModule,
     NimbleMenuItemModule, NimbleCheckboxModule, NimbleToggleButtonModule, NimbleBreadcrumbModule, NimbleBreadcrumbItemModule,
     NimbleIconAddModule, NimbleSwitchModule, NimbleToolbarModule, NimbleMenuButtonModule, NimbleComboboxModule, NimbleTooltipModule,
-    NimbleCardButtonModule, NimbleDialogModule, NimbleRadioGroupModule, NimbleRadioModule } from '@ni/nimble-angular';
+    NimbleCardButtonModule, NimbleDialogModule, NimbleRadioGroupModule, NimbleRadioModule, NimbleSpinnerModule } from '@ni/nimble-angular';
+import { NimbleTableModule } from 'projects/ni/nimble-angular/src/public-api';
 import { AppComponent } from './app.component';
 import { CustomAppComponent } from './customapp/customapp.component';
 import { HeaderComponent } from './header/header.component';
@@ -43,6 +44,7 @@ import { HeaderComponent } from './header/header.component';
         NimbleBreadcrumbModule,
         NimbleBreadcrumbItemModule,
         NimbleIconAddModule,
+        NimbleSpinnerModule,
         NimbleSwitchModule,
         NimbleToolbarModule,
         NimbleComboboxModule,
@@ -52,6 +54,7 @@ import { HeaderComponent } from './header/header.component';
         NimbleDialogModule,
         NimbleRadioGroupModule,
         NimbleRadioModule,
+        NimbleTableModule,
         RouterModule.forRoot([
             { path: '', redirectTo: '/customapp', pathMatch: 'full' },
             { path: 'customapp', component: CustomAppComponent }