+++ packages/nimble-components/src/all-components.ts
@@ -24,6 +24,7 @@ import './number-field';
 import './radio';
 import './radio-group';
 import './select';
+import './spinner';
 import './switch';
 import './tab';
 import './tab-panel';