+++ packages/nimble-components/src/wafer-map/tests/wafer-map.stories.ts
@@ -96,22 +96,28 @@ const metadata: Meta<WaferMapArgs> = {
             WARNING - The wafermap is still in development and considered
             experimental. It is not recommended for application use.
         </div>
-        <nimble-wafer-map
-            colors-scale-mode="${x => x.colorScaleMode}"
-            ?die-labels-hidden="${x => x.dieLabelsHidden}"
-            die-labels-suffix="${x => x.dieLabelsSuffix}"
-            max-characters="${x => x.maxCharacters}"
-            orientation="${x => x.orientation}"
-            quadrant="${x => x.quadrant}"
-            :colorScale="${x => x.colorScale}"
-            :dies="${x => getDiesSet(x.dies, wafermapDieSets)}"
-            :highlightedValues="${x => getHighLightedValueSets(
+        <div class="wrapper">
+            <nimble-wafer-map
+                colors-scale-mode="${x => x.colorScaleMode}"
+                ?die-labels-hidden="${x => x.dieLabelsHidden}"
+                die-labels-suffix="${x => x.dieLabelsSuffix}"
+                max-characters="${x => x.maxCharacters}"
+                orientation="${x => x.orientation}"
+                quadrant="${x => x.quadrant}"
+                :colorScale="${x => x.colorScale}"
+                :dies="${x => getDiesSet(x.dies, wafermapDieSets)}"
+                :highlightedValues="${x => getHighLightedValueSets(
         x.highlightedValues,
         highLightedValueSets
     )}"
-        >
-        </nimble-wafer-map>
+            >
+            </nimble-wafer-map>
+        </div>
         <style class="code-hide">
+            .wrapper {
+                width: 50vw;
+                height: 50vh;
+            }
             #usage-warning {
                 color: red;
                 font: var(${bodyFont.cssCustomProperty});